requirements.txt
chainlit==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
python-dotenv==1.0.0
numpy==1.24.3